# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервиса аутентификации для платформы SwiftTalk",
        "title": "SwiftTalk Auth Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/authService/",
    "paths": {
        "/login": {
            "post": {
                "description": "Аутентифицирует пользователя и возвращает токены доступа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход в систему",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bff.LoginPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "schema": {
                            "$ref": "#/definitions/bff.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка в параметрах запроса",
                        "schema": {
                            "$ref": "#/definitions/bff.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/bff.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "description": "Обновляет access и refresh токены",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление токена",
                "responses": {
                    "200": {
                        "description": "Успешное обновление токенов",
                        "schema": {
                            "$ref": "#/definitions/bff.TokenResponse"
                        }
                    },
                    "401": {
                        "description": "Невалидный refresh token",
                        "schema": {
                            "$ref": "#/definitions/bff.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/bff.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Регистрирует нового пользователя и возвращает токены доступа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bff.RegisterPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная регистрация",
                        "schema": {
                            "$ref": "#/definitions/bff.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка в параметрах запроса",
                        "schema": {
                            "$ref": "#/definitions/bff.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/bff.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/verify": {
            "get": {
                "description": "Проверяет валидность access token и возвращает информацию о пользователе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Проверка токена",
                "responses": {
                    "200": {
                        "description": "Успешная проверка токена",
                        "schema": {
                            "$ref": "#/definitions/bff.VerifyTokenResponse"
                        }
                    },
                    "401": {
                        "description": "Невалидный access token",
                        "schema": {
                            "$ref": "#/definitions/bff.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/bff.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "bff.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "bff.LoginPayload": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "bff.RegisterPayload": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "bff.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "bff.VerifyTokenResponse": {
            "type": "object",
            "properties": {
                "is_valid": {
                    "type": "boolean"
                },
                "user_uuid": {
                    "type": "string"
                }
            }
        }
    }
}